% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confinterpret.R
\name{confinterpret}
\alias{confinterpret}
\title{Descriptive interpretations of confidence intervals.}
\usage{
confinterpret(ci, interpretation_set, boundaries, comparison_labels = NULL)
}
\arguments{
\item{ci}{A single row from a matrix of the type returned by \code{confint()},
containing the confidence interval for the parameter estimate. The two
columns provide the lower and upper confidence limits.}

\item{interpretation_set}{List-based object that specifies the boundaries between regions that each
of the confidence limits can fill in, and the interpretations to be
returned in each of the cases.}

\item{boundaries}{Vector of numbers specifying the values for each of the boundaries defined
in the \code{interpretation_set}.}

\item{comparison_labels}{Character vector specifying the labels to be used within the
interpretation to describe the comparison. Required if the
\code{interpretation_set} includes a $placeholders entry. Null otherwise.}
}
\value{
A list object with elements stating the interpretation in different
  formats.
}
\description{
Produces descriptive interpretations of confidence intervals, depending
on the type of test specified by an \code{\link{interpretation_set}}.
}
\details{
Helpful wrapper functions are provided for some commonly used types of
test:

\describe{
  \item{Superiority tests}{\code{\link{interpret_superiority}}}
  \item{Non-inferiority tests}{\code{\link{interpret_noninferiority}}}
  \item{Equivalence tests}{\code{\link{interpret_equivalence}}}
}
}
\examples{
# Establish a test confidence interval
ci_test <- matrix(c(-0.1,0.1),
                  nrow = 1, dimnames = list("estimate",
                                            c("2.5 \%","97.5 \%")))
confinterpret(ci_test, interpretations_superiority, 0,
  comparison_labels = c(comparison_intervention = "Treatment as usual",
                        tested_intervention = "New treatment"))

}
