% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noninferiority.R
\docType{data}
\name{interpretations_noninferiority}
\alias{interpretations_noninferiority}
\title{Interpretation set for non-inferiority tests}
\format{An object of class \code{interpretation_set} of length 3.}
\usage{
interpretations_noninferiority
}
\description{
An \code{\link{interpretation_set}} object used for conducting
non-inferiority tests. A convenient wrapper function,
\code{\link{interpret_noninferiority}}, is provided, making use of
this object.
}
\keyword{datasets}
