% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_intervals}
\alias{plot_intervals}
\title{Plot intervals}
\usage{
plot_intervals(intervals, estimates = NULL, interval_type = "norm", ...)
}
\arguments{
\item{intervals}{The interval(s) to be plotted. Two column matrix.}

\item{estimates}{Estimates for each of the intervals (optional).}

\item{interval_type}{Set the way the interval is presented. Current options are
\code{c("norm", "unif")} for a normal distribution-based curve
and a box, respectively.}
}
\description{
Plot intervals on a canvas, typically prepared with
\code{plot_region_canvas()}.
}
