% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_intervals_norm}
\alias{plot_intervals_norm}
\title{Plot intervals as curved (normal distribution) areas}
\usage{
plot_intervals_norm(intervals, estimates = NULL, y_scale = 1,
  interval_value_labels = FALSE, estimate_value_labels = FALSE,
  interval_labels_offset = c(0, 0, 0.2, 0.2), estimate_labels_offset = c(0,
  0.6), ...)
}
\description{
Plot intervals as curved (normal distribution) areas
}
