% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{strwidthl}
\alias{strwidthl}
\title{Obtain string widths in (approximate) multiple of lines.}
\usage{
strwidthl(s)
}
\arguments{
\item{s}{A character vector whose width is to be determined.}
}
\description{
Obtain string widths in (approximate) multiple of lines.
}
