% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_result.R
\name{validate_interpretation_result}
\alias{validate_interpretation_result}
\title{Validator for interpretation_result objects}
\usage{
validate_interpretation_result(x)
}
\arguments{
\item{x}{An object to be checked to see whether it is a valid
interpretation_result.}
}
\value{
The interpretation_result object that was input, if no errors are
  found.
}
\description{
Checks some features of the passed object to see whether they are as
expected for the class. See \code{\link{interpretation_result}}
documentation for definition of the class.
}
