% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_set.R
\name{validate_interpretation_set}
\alias{validate_interpretation_set}
\title{Validator for interpretation_set objects}
\usage{
validate_interpretation_set(interpretation_set)
}
\arguments{
\item{interpretation_set}{An object to be checked to see whether it is a valid interpretation_set.}
}
\value{
The interpretation_set object that was input, if no errors are
  found.
}
\description{
Checks some features of the passed object to see whether they are as
expected for the class. See \code{\link{interpretation_set}} documentation
for definition of the class.
}
