% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_result_set.R
\name{validate_result_set}
\alias{validate_result_set}
\title{Validates a set of interpretation_result objects}
\usage{
validate_result_set(x)
}
\arguments{
\item{x}{A list of \code{interpretation_result} objects, length at least 2. The
objects may optionally be named.}
}
\description{
Checks that a set of interpretation_result objects has been correctly
assembled for use in the plotting function.
}
