% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalence.R
\docType{data}
\name{interpretations_equivalence}
\alias{interpretations_equivalence}
\title{Interpretation set for equivalence tests}
\format{An object of class \code{interpretation_set} of length 3.}
\usage{
interpretations_equivalence
}
\description{
An \code{\link{interpretation_set}} object used for conducting equivalence
tests. A convenient wrapper function, \code{\link{interpret_equivalence}},
is provided, making use of this object.
}
\keyword{datasets}
