% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noninferiority.R
\docType{data}
\name{interpretations_noninferiority}
\alias{interpretations_noninferiority}
\title{Interpretation set for non-inferiority tests}
\format{An object of class \code{interpretation_set} of length 3.}
\usage{
interpretations_noninferiority
}
\description{
An \code{\link{interpretation_set}} object used for conducting
non-inferiority tests. A convenient wrapper function,
\code{\link{interpret_noninferiority}}, is provided, making use of
this object.
}
\details{
This \code{interpretation_set} object contains \code{placeholders} for
descriptive names of the comparison intervention and tested intervention.
When used with \code{\link{confinterpret}} these are provided via the
\code{comparison_labels} parameter as a named character vector of length 2,
\code{c(comparison_intervention = "Your control / comparison intervention",
tested_intervention = "Your new / tested intervention")}. When using the
convenience wrapper function, this is handled through the \code{groups}
parameter.
}
\keyword{datasets}
