% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superiority.R
\docType{data}
\name{interpretations_superiority}
\alias{interpretations_superiority}
\title{Interpretation set for superiority tests}
\format{An object of class \code{interpretation_set} of length 3.}
\usage{
interpretations_superiority
}
\description{
An \code{\link{interpretation_set}} object used for conducting superiority
tests. A convenient wrapper function, \code{\link{interpret_superiority}},
is provided, making use of this object.
}
\keyword{datasets}
