% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{label_ontop_boundaries}
\alias{label_ontop_boundaries}
\title{Label the boundaries on top of the plot.}
\usage{
label_ontop_boundaries(boundaries, extra_boundaries = NULL)
}
\description{
If plotting values or ranges may want to call this directly last
to ensure it is on top, and specify no labels in the canvas plotting call.
}
