% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_intervals_unif}
\alias{plot_intervals_unif}
\title{Plot intervals as uniform (box) areas}
\usage{
plot_intervals_unif(intervals, estimates = NULL,
  interval_value_labels = FALSE, estimate_value_labels = FALSE,
  interval_labels_offset = c(-0.1, 0.1, 0, 0), estimate_labels_offset = c(0,
  0.6), ...)
}
\arguments{
\item{interval_labels_offset}{Amount to offset interval labels by from the centre of the end
of the interval's plot. \code{c(x1, x2, y1, y2)}.}

\item{estimate_labels_offset}{Amount to offset estimate labels by. \code{c(x, y)}.}
}
\description{
Plot intervals as uniform (box) areas
}
