% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_intervals_unif}
\alias{plot_intervals_unif}
\title{Plot intervals as uniform (box) areas}
\usage{
plot_intervals_unif(intervals, estimates = NULL,
  interval_value_labels = FALSE, estimate_value_labels = FALSE,
  interval_labels_offset = c(0, 0, 0.3, 0.3), estimate_labels_offset = c(0,
  0.3), ...)
}
\arguments{
\item{intervals}{The interval(s) to be plotted. Two column matrix.}

\item{estimates}{Estimates for each of the intervals (optional).}

\item{interval_value_labels}{Logical value specifying whether interval value labels are to be added.}

\item{estimate_value_labels}{Logical value specifying whether estimate value labels are to be added.}

\item{...}{Further parameters to be passed on.}
}
\description{
Plot intervals as uniform (box) areas
}
